add ai update

+++ services/updater.cpp
@@ -61,10 +61,10 @@ using Updater::Utils::Trim;
 using namespace Hpackage;
 using namespace Utils;
 
-int g_percentage = 100;
-int g_tmpProgressValue;
-int g_tmpValue;
-bool g_canceled = false;
+thread_local int g_percentage = 100;
+thread_local int g_tmpProgressValue;
+thread_local int g_tmpValue;
+thread_local bool g_canceled = false;
  
 void SetCancelStatus(bool isCanceled)
 {
@@ -624,14 +624,15 @@ void ExcuteSubProc(const UpdaterParams &upParams, const std::string &fullPath, i
         }
     }
     const std::string retryPara = upParams.retryCount > 0 ? "retry=1" : "retry=0";
+    const std::string procName = 
     int ret = -1;
     if (upParams.updateBin.size() > 0) {
         LOG(INFO) << "Binary Path:" << upParams.updateBin[upParams.pkgLocation].c_str();
-        ret = execl(fullPath.c_str(), fullPath.c_str(), upParams.updateBin[upParams.pkgLocation].c_str(),
+        ret = execl(fullPath.c_str(), procName.c_str(), upParams.updateBin[upParams.pkgLocation].c_str(),
             std::to_string(pipeWrite).c_str(), retryPara.c_str(), nullptr);
     } else if (upParams.updatePackage.size() > 0) {
         LOG(INFO) << "Binary Path:" << packagePath.c_str();
-        ret = execl(fullPath.c_str(), fullPath.c_str(), packagePath.c_str(),
+        ret = execl(fullPath.c_str(), procName.c_str(), packagePath.c_str(),
             std::to_string(pipeWrite).c_str(), retryPara.c_str(), nullptr);
     }
     if (ret < 0) {
@@ -727,6 +728,9 @@ static std::string GetBinaryPathFromBin(PkgManager::PkgManagerPtr pkgManager, Up
 
 static std::string GetBinaryPath(PkgManager::PkgManagerPtr pkgManager, UpdaterParams &upParams)
 {
+    if (upParams.) {
+        return "/bin/updater_binary";
+    }
     std::string fullPath = GetWorkPath() + std::string(UPDATER_BINARY);
     (void)Utils::DeleteFile(fullPath);
     std::string packagePath = GetCurrentPackagePath(upParams);
